Begun adding protocol tests.
