+++ propertyestimator/tests/test_workflow/test_protocols.py
@@ -231,217 +231,3 @@ def test_base_simulation_protocols():
 
         result = extract_uncorrelated_statistics.execute(temporary_directory, ComputeResources())
         assert not isinstance(result, PropertyEstimatorException)
-
-
-def test_addition_subtract_protocols():
-
-    with tempfile.TemporaryDirectory() as temporary_directory:
-
-        quantity_a = EstimatedQuantity(1*unit.kelvin, 0.1*unit.kelvin, 'dummy_source_1')
-        quantity_b = EstimatedQuantity(2*unit.kelvin, 0.2*unit.kelvin, 'dummy_source_2')
-
-        add_quantities = AddValues('add')
-        add_quantities.values = [quantity_a, quantity_b]
-
-        result = add_quantities.execute(temporary_directory, ComputeResources())
-
-        assert not isinstance(result, PropertyEstimatorException)
-        assert add_quantities.result.value == 3 * unit.kelvin
-
-        sub_quantities = SubtractValues('sub')
-        sub_quantities.value_b = quantity_b
-        sub_quantities.value_a = quantity_a
-
-        result = sub_quantities.execute(temporary_directory, ComputeResources())
-
-        assert not isinstance(result, PropertyEstimatorException)
-        assert sub_quantities.result.value == 1 * unit.kelvin
-
-
-@pytest.mark.parametrize("filter_role", [Substance.ComponentRole.Solute,
-                                         Substance.ComponentRole.Solvent,
-                                         Substance.ComponentRole.Ligand,
-                                         Substance.ComponentRole.Receptor])
-def test_substance_filtering_protocol(filter_role):
-    """Tests that the protocol to filter substances by
-    role correctly works."""
-
-    def create_substance():
-
-        test_substance = Substance()
-
-        test_substance.add_component(Substance.Component('C', role=Substance.ComponentRole.Solute),
-                                     Substance.ExactAmount(1))
-
-        test_substance.add_component(Substance.Component('CC', role=Substance.ComponentRole.Ligand),
-                                     Substance.ExactAmount(1))
-
-        test_substance.add_component(Substance.Component('CCC', role=Substance.ComponentRole.Receptor),
-                                     Substance.ExactAmount(1))
-
-        test_substance.add_component(Substance.Component('O', role=Substance.ComponentRole.Solvent),
-                                     Substance.MoleFraction(1.0))
-
-        return test_substance
-
-    filter_protocol = FilterSubstanceByRole('filter_protocol')
-    filter_protocol.input_substance = create_substance()
-
-    filter_protocol.component_role = filter_role
-    filter_protocol.execute('', ComputeResources())
-
-    assert len(filter_protocol.filtered_substance.components) == 1
-    assert filter_protocol.filtered_substance.components[0].role == filter_role
-
-
-def _build_input_output_substances():
-    """Builds sets if input and expected substances for the
-    `test_build_coordinate_composition` test.
-
-    Returns
-    -------
-    list of tuple of Substance and Substance
-        A list of input and expected substances.
-    """
-
-    # Start with some easy cases
-    substances = [
-        (Substance.from_components('O'), Substance.from_components('O')),
-        (Substance.from_components('O', 'C'), Substance.from_components('O', 'C')),
-        (Substance.from_components('O', 'C', 'CO'), Substance.from_components('O', 'C', 'CO'))
-    ]
-
-    # Handle some cases where rounding will need to occur.
-    input_substance = Substance()
-    input_substance.add_component(Substance.Component('O'), Substance.MoleFraction(0.41))
-    input_substance.add_component(Substance.Component('C'), Substance.MoleFraction(0.59))
-
-    expected_substance = Substance()
-    expected_substance.add_component(Substance.Component('O'), Substance.MoleFraction(0.4))
-    expected_substance.add_component(Substance.Component('C'), Substance.MoleFraction(0.6))
-
-    substances.append((input_substance, expected_substance))
-
-    input_substance = Substance()
-    input_substance.add_component(Substance.Component('O'), Substance.MoleFraction(0.59))
-    input_substance.add_component(Substance.Component('C'), Substance.MoleFraction(0.41))
-
-    expected_substance = Substance()
-    expected_substance.add_component(Substance.Component('O'), Substance.MoleFraction(0.6))
-    expected_substance.add_component(Substance.Component('C'), Substance.MoleFraction(0.4))
-
-    substances.append((input_substance, expected_substance))
-
-    return substances
-
-
-@pytest.mark.parametrize("input_substance, expected", _build_input_output_substances())
-def test_build_coordinate_composition(input_substance, expected):
-    """Tests that the build coordinate protocols correctly report
-    the composition of the built system."""
-
-    build_coordinates = BuildCoordinatesPackmol('build_coordinates')
-    build_coordinates.max_molecules = 10
-    build_coordinates.substance = input_substance
-
-    with tempfile.TemporaryDirectory() as directory:
-        assert not isinstance(build_coordinates.execute(directory, None), PropertyEstimatorException)
-
-    assert build_coordinates.output_substance == expected
-
-
-def test_solvation_protocol():
-    """Tests solvating a single methanol molecule in water."""
-
-    methanol_substance = Substance()
-    methanol_substance.add_component(Substance.Component('CO'), Substance.ExactAmount(1))
-
-    water_substance = Substance()
-    water_substance.add_component(Substance.Component('O'), Substance.MoleFraction(1.0))
-
-    with tempfile.TemporaryDirectory() as temporary_directory:
-
-        build_methanol_coordinates = BuildCoordinatesPackmol('build_methanol')
-
-        build_methanol_coordinates.max_molecules = 1
-        build_methanol_coordinates.substance = methanol_substance
-
-        build_methanol_coordinates.execute(temporary_directory, ComputeResources())
-
-        solvate_coordinates = SolvateExistingStructure('solvate_methanol')
-
-        solvate_coordinates.max_molecules = 9
-        solvate_coordinates.substance = water_substance
-        solvate_coordinates.solute_coordinate_file = build_methanol_coordinates.coordinate_file_path
-
-        solvate_coordinates.execute(temporary_directory, ComputeResources())
-
-        solvated_pdb = PDBFile(solvate_coordinates.coordinate_file_path)
-
-        assert solvated_pdb.topology.getNumResidues() == 10
-
-
-@pytest.mark.parametrize('charge_backend', [BuildTLeapSystem.ChargeBackend.OpenEye,
-                                            BuildTLeapSystem.ChargeBackend.AmberTools])
-def test_topology_mol_to_mol2(charge_backend):
-    """Tests taking an openforcefield topology molecule, generating a conformer,
-    calculating partial charges, and writing it to mol2."""
-
-    # Testing to find the correct connectivity information should indicate 
-    # that the mol2 conversion was successful
-    expected_contents = ['''
-@<TRIPOS>BOND
-     1    1    3 1
-     2    2    3 1
-     3    3    9 1
-     4    3    8 1
-     5    1    4 1
-     6    2    5 1
-     7    2    6 1
-     8    2    7 1''', '''@<TRIPOS>BOND
-     1    4    6 ar
-     2    2    4 ar
-     3    1    2 ar
-     4    1    3 ar
-     5    3    5 ar
-     6    5    6 ar
-     7    6    7 1
-     8    4   11 1
-     9    2    9 1
-    10    1    8 1
-    11    3   10 1
-    12    5   12 1
-    13    7   13 1
-    14    7   14 1
-    15    7   15 1''']
-
-    # Constructing the molecule using this SMILES will ensure that the reference molecule
-    # (generated here) and topology molecule (generated below from PDB) have a different atom order
-    ethanol_smiles = 'C(O)C'
-    toluene_smiles = 'c1ccccc1C'
-    
-    ethanol = Molecule.from_smiles(ethanol_smiles)
-    toluene = Molecule.from_smiles(toluene_smiles)
-
-    pdb_file = PDBFile(get_data_filename('test/molecules/ethanol_toluene.pdb'))
-    topology = Topology.from_openmm(pdb_file.topology, unique_molecules=[ethanol, toluene])
-
-    for topology_molecule_index, topology_molecule in enumerate(topology.topology_molecules):
-
-        with NamedTemporaryFile(suffix='.mol2') as output_file:
-
-            BuildTLeapSystem._topology_molecule_to_mol2(topology_molecule,
-                                                        output_file.name,
-                                                        charge_backend=charge_backend)
-
-            mol2_contents = open(output_file.name).read()
-
-            # Ensure we find the correct connectivity info
-            assert expected_contents[topology_molecule_index] in mol2_contents
-
-            # Ensure that the first atom has nonzero coords and charge
-            first_atom_line = mol2_contents.split('\n')[7].split()
-            assert float(first_atom_line[2]) != 0.
-            assert float(first_atom_line[3]) != 0.
-            assert float(first_atom_line[4]) != 0.
-            assert float(first_atom_line[8]) != 0.
